+++ packages/expo-image/src/Image.types.ts
@@ -8,8 +8,7 @@ export type ImageSource = {
   uri?: string;
   /**
    * An object representing the HTTP headers to send along with the request for a remote image.
-   * @platform android
-   * @platform ios
+   * On web requires the `Access-Control-Allow-Origin` header returned by the server to include the current domain.
    */
   headers?: Record<string, string>;
   /**